+++ js/main.js
@@ -1,13 +1,12 @@
-import {blockForm, unblockFilters, initValidation} from './form.js';
+import {blockForm, unblockFilters, initValidation, unblockForm} from './form.js';
 import {initMap, createAdsMarkers} from './map.js';
 import {getData} from './api.js';
-import {showMessageError} from './util.js';
+import {showMessageError} from './popup.js';
 
 blockForm();
 initValidation();
-initMap();
 
-getData((ads) => {
+initMap(() => getData((ads) => {
+  unblockFilters();
   createAdsMarkers(ads);
-}, showMessageError, unblockFilters);
-
+}, showMessageError));
